Some fixes in auth actions and Login Page

+++ app/containers/Auth/actions.js
@@ -12,14 +12,13 @@ import { SET_USER_DATA, LOADING_USER_DATA, USER_LOGOUT } from './constants';
 export const loginUser = (email, password) => async dispatch => {
   dispatch(authStart());
   try {
-    const { data } = api.post('/user/login/admin', {
+    const { data } = await api.post('/user/login/admin', {
       email,
       password,
     });
     dispatch(authSuccess(data));
   } catch (error) {
-    console.log(error.response);
-    throw error;
+    throw error.response.status;
   }
 };
 
@@ -28,12 +27,15 @@ export const authStart = () => ({
 });
 
 export const authSuccess = session => {
-  const expiration = new Date();
-  expiration.setDate(expiration.getDate() + 1);
-  Cache.setItem('token', session.token, {
-    expires: expiration.getTime(),
-    priority: 1,
-  });
+  const token = Cache.getItem('token');
+  if (!token) {
+    const expiration = new Date();
+    expiration.setDate(expiration.getDate() + 1);
+    Cache.setItem('token', session.token, {
+      expires: expiration.getTime(),
+      priority: 1,
+    });
+  }
   return {
     type: SET_USER_DATA,
     payload: {
@@ -48,20 +50,22 @@ export const authFail = () => ({
 
 export const logout = () => async dispatch => {
   const token = Cache.getItem('token');
-  try {
-    await api.delete('/user/me/token', {
-      headers: {
-        Authorization: `Bearer ${token}`,
-      },
-    });
-    dispatch({
-      type: USER_LOGOUT,
-    });
-  } catch (error) {
-    console.log(error.response);
-    throw new Error(
-      'Access Token missing or Not authorized to perform this action',
-    );
+  if (token) {
+    try {
+      await api.delete('/user/me/token', {
+        headers: {
+          Authorization: `Bearer ${token}`,
+        },
+      });
+      dispatch({
+        type: USER_LOGOUT,
+      });
+      Cache.removeItem('token');
+    } catch (error) {
+      throw new Error(
+        'Access Token missing or Not authorized to perform this action',
+      );
+    }
   }
 };
 
@@ -77,7 +81,6 @@ export const checkSession = () => async dispatch => {
       });
       dispatch(authSuccess(data));
     } catch (error) {
-      console.log(error.response);
       dispatch(authFail());
       throw new Error(
         'Access Token missing or Not authorized to perform this action',
